# This is a combination of 2 commits. # The first commit's message is:
# This is a combination of 4 commits.
# The first commit's message is:
# This is a combination of 3 commits.
# The first commit's message is:
[SW-158] Added Dataset to H2O spark interface.

[SW-158] Added Dataset to H2O spark interface.

A couple of tests were failing.
One was when the column name in receiver was different from the one in the data.
Fixed it by allowing to rename single-column data.
Another was that python tests were _actually_ failing (the failures are not reported, that's a bug, kinda). So had to tweak all those implicit sqlContexts passing. Not it works.

Rewrote the Dataset functionality to work with "Converters" - preliminary version.
Seems like all the tests pass. Have to refactor stuff now.

[SW-158] Added Dataset to H2O spark interface.

A couple of tests were failing.
One was when the column name in receiver was different from the one in the data.
Fixed it by allowing to rename single-column data.
Another was that python tests were _actually_ failing (the failures are not reported, that's a bug, kinda). So had to tweak all those implicit sqlContexts passing. Not it works.

Rewrote the Dataset functionality to work with "Converters" - preliminary version.
Seems like all the tests pass. Have to refactor stuff now.

# This is the 2nd commit message:

A couple of tests were failing.
One was when the column name in receiver was different from the one in the data.
Fixed it by allowing to rename single-column data.
Another was that python tests were _actually_ failing (the failures are not reported, that's a bug, kinda). So had to tweak all those implicit sqlContexts passing. Not it works.

# This is the 3rd commit message:

Rewrote the Dataset functionality to work with "Converters" - preliminary version.
Seems like all the tests pass. Have to refactor stuff now.

# This is the 2nd commit message:

A couple of tests were failing.
One was when the column name in receiver was different from the one in the data.
Fixed it by allowing to rename single-column data.
Another was that python tests were _actually_ failing (the failures are not reported, that's a bug, kinda). So had to tweak all those implicit sqlContexts passing. Not it works.

# This is the 3rd commit message:

Rewrote the Dataset functionality to work with "Converters" - preliminary version.
Seems like all the tests pass. Have to refactor stuff now.

# This is the 4th commit message:

Removed duplicate methods.

# This is the 2nd commit message:

Restored some of Yakub's new functionality